requirements.txt
absl-py==2.3.1
immutabledict==4.2.2
numpy==2.4.1
ortools==9.15.6755
pandas==3.0.0
protobuf==6.33.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
typing_extensions==4.15.0
